Fix Host stuck at the right side
Cause:
This bug only repro on chrome v33 and v34

<div class="vertically-centered">
  <div class="horizontally-centered">
    <div id="video-container">
      <video id="mediasource-video-output"></video>
      <div id="client-plugin-container"></div>
    </div>
  </div>
</div>

Due to a chrome layout bug, the width of video-container collapses to 0 when the zoom level is greater than 100%. As a result, the flexbox defined in horizonally-centered fails to center the element.

Fix:
Set width=100% on video-container so that its width remains valid.  Since video-container now sizes to the parent, we need to move the horizontal-centered flexbox to the children of video-container.

BUG=357114

Review URL: https://codereview.chromium.org/217363002