Cast: Clarify the unit of bitrate
This change is to clarify the unit used in cast.streaming APIs for
bitrate. It should be in kilobits per second. And the unit conversion
should be done just before converting to VideoSenderConfig and
AudioSenderConfig which use bits per second as unit.

This will improve the video quality for existing extensions which use
kilobits per second.

BUG=354267
NOTRY=true

Review URL: https://codereview.chromium.org/214663008