Reduce the scope of CGL current contexts
There are a few crash reports coming in where the destructor
~ScopedCGLSetCurrentContext is being implicated inside
CompositorSwapBuffers. Reduce the scope of the 
ScopedCGLSetCurrentContext in CompositorSwapBuffers.

At a minimum, this will make the crash reports reveal which
place is causing problems. At a maximum, this may fix the
issue (in particular, LayoutLayers may call displayIfNeeded,
which may mess with the current context, potentially causing
crashes, as ScopedCGLSetCurrentContext does not expect
this).

BUG=245900
NOTRY=true

Review URL: https://codereview.chromium.org/216733002