Android: Fix null ptr deref
With r258476 it became possible to call RunAckCallbacks() after
host_ was reset to NULL from Destroy().

ReleaseLocksOnSurface() can call InternalSwapCompositorFrame()
which queues an ACK if we are visible.

Make sure that when we call SetContentViewCore(NULL) the first
time while |host_| is still valid, we are not left with an unsent
ACK.

BUG=278466
NOTRY=True

Review URL: https://codereview.chromium.org/212123011