Handle direction control in compositor Animations
This patch is the compositor side. The blink side is handled in issue #178263006

Handle reverse animations in the compositor. Currently, compositor only
understand alternate direction. And in order to reverse an animation, it
needs to be passed a reversed timing function.

This patch introduces a Direction enum parameter in animation.

BUG=348071

Review URL: https://codereview.chromium.org/180153010