TaskManager WebContentsInformation: ExtensionInformation (second attempt)
This is a re-land of https://codereview.chromium.org/197883024 with
bugs fixed. What's different are the calls to GetExtensionForWebContents,
particularly the call done by CheckOwnership(). This verifies
that the WebContents is in fact an chrome-extension:// WebContents.
This logic now matches the filtering done by
NOTIFICATION_EXTENSION_VIEW_REGISTERED in the original code.

BUG=356087,355501,355642,348836
TEST=browser_tests,manual verification of feedback extension system information in official build

Review URL: https://codereview.chromium.org/217463002