Add a function to create a bookmark app from a WebApplicationInfo.
This CL adds a function which installs a bookmark app that is created
from a WebApplicationInfo. This is part of an effort to consolidate
code that deals with bookmark apps and write tests for them and is part of
laying the groundwork for syncing of bookmark apps.

BUG=318607

Review URL: https://codereview.chromium.org/202523011