Add base.gyp:base_static as dependency of nacl component
The switches defined in base_switches.h are used in
nacl_process_host.cc, add dependency is clearly needed for other NaCl
embedder.

This dependency is added in chrome/chrome_common.gypi, thus this issue
is hidden for chrome build.

This is part of an effort to componentize NaCl code.

BUG=244791

Review URL: https://codereview.chromium.org/218763003