ui: Replace PostedSwapQueue with a SwapState enum.
The posted swap queue can only ever have a single swap in it, which
is verified by DCHECKs, and it only has a single swap type in the
enum, so this deque is pointless and it is causing mysterious crashes.

Refactor the posted swap queue into a single SwapState enum to track
if a swap was/is currently posted or not.

R=piman@chromium.org
BUG=348609

Review URL: https://codereview.chromium.org/214013002