Refactor VideoDestinationHandler to implement MediaStreamVideoSource. VideoDestinationHandler is moved to media/webrtc since it needs to use libyuv. It also make sence since its a private API only used for WebRtc.
This cl also remove the WebRtcVideoCaptureAdapter::SetCurrrentFormat and make sure video resolution constraints are not passed to libJingle. Libjingle do not need to know the capture resolution or the used resolution constraints and that is infact how
VideoDestionationHandler can work today.
This is needed to simplify the adapter and the use of MediaStreamVideoSource since its currently not possible to call WebRtcVideoCaptureAdapter::SetCurrrentFormat before the MediaStreamVideoSource::OnSupportedFormat has been called.

BUG=334243

// TBR Jam for gyp changes, yzshen for the header file path change in Pepper.

R=joi@chromium.org, ronghuawu@chromium.org
TBR=jam, yzshen1

Review URL: https://codereview.chromium.org/212973002