Fire accessibility focus/blur events on root web document.
When the aura Window containing the RWHVA gets or loses focus,
we must fire the corresponding accessibility event on that
object, which corresponds to the root "document" in the web
hierarchy. This worked before the Aura switch because native
HWNDs fired these events automatically.

BUG=347448

Review URL: https://codereview.chromium.org/214243005