Add an 'obscured' flag to views::Label.
For the password management bubble, we're rendering passwords into a
label after replacing the string with bullets. This is something that's
available at the platform level on OSX (NSSecureTextField); it seems
reasonable to add the capability to labels themselves, rather than
special-casing it into a view just for this bubble.

'git cl format' produced minor formatting changes in the effected lines.

BUG=328339
R=msw@chromium.org, sadrul@chromium.org

Review URL: https://codereview.chromium.org/222033002