Create a gsutil api function that takes a URL
GSUtilApi has an upload step that gives an https:// URL to the upload.
It also has a download step that accepts a gs:// URL to the file.  It
is convenient to also be able to download with the https:// URL provided
by the upload step.

BUG=358659
R=agable@chromium.org, mrapisarda@google.com

Review URL: https://codereview.chromium.org/221313002