Remove WebTestProxy repaint mask
This repaint mask used to display a large partially transparent black
layer over the whole page to try to visually display the parts of the
page that have changed.

Most layout tests have had display() removed from them.  The only tests
that still call testRunner.display() directly or indirectly are:

    fast/canvas/webgl/webgl-composite-modes-repaint.html
    inspector/timeline/timeline-decode-resize.html
    inspector/timeline/timeline-paint.html
    inspector/timeline/timeline-recalculate-styles.html
    plugins/windowless_plugin_paint_test.html

Of those tests, only the first one is a pixel test, but the repaint mask
is not visible (because the final frame repaints everything).
Therefore, this can be removed safely with no changes in Blink.

R=abarth@chromium.org
BUG=345027

Review URL: https://codereview.chromium.org/224253005