Merge w3c test expectations from LayoutTests/TestExpectations .
Some of the existing test expectations refer to files that have been
removed in the upstream W3C repo, and hence will cause lint-test-expectations
to fail on the next W3C roll. If we move these expectations into the
downstream test_expectations_w3c.txt file, we can then update the expectations
atomically in the roll and avoid any issues.

Unfortunately, this requires us to land five CLs as a result and hope
that nobody needs to adjust the Blink TestExpectations in the meantime :(.

TBR=abarth

Review URL: https://codereview.chromium.org/226573002