Relanding this change. The bug 332430 which caused this to be reverted will be addressed in an upcoming change.
Invoke cursor-hiding logic on touch for Windows

Continue to invoke cursor-hiding logic on touch
for Windows even though the OS hides the cursor
for us. The reason is that this code path is
needed in order to call
RenderWidgetHostViewAura::
NotifyRendererOfCursorVisibilityState()
in order for the renderer to know to suppress
generation of synthetic mousemove events
while touch scrolling. This avoids unwanted
hover effects from being invoked when the
cursor is not visible.

To avoid regressions in behavior (such as
the one mentioned in
https://codereview.chromium.org/56053004),
make the cursor state global among all
root windows. This is done in
https://codereview.chromium.org/111043002/.

BUG=322250
TEST=To be added later (crbug.com/328549)
R=sky@chromium.org, tdanderson@chromium.org

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=240976

Review URL: https://codereview.chromium.org/221933014