Encrypted Media: Drop wrong DCHECK in RendererMediaPlayerManager.
Now the |player_id| is generated in
RendererMediaPlayerManager::RegisterMediaPlayer() and the |cdm_id| is generated
in ProxyMediaKeys::ProxyMediaKeys(). They are different in general and can be
the same only be accident. This DCHECK is wrong which assumes the |cdm_id| is
always the same as the |player_id|.

BUG=338910

Review URL: https://codereview.chromium.org/226043006