linux: Make sure event-dispatcher gets removed correctly.
A DesktopDragDropClientAuraX11 owns an instance of X11DragContext, which gets
destroyed when the DesktopDragDropClientAuraX11 gets destroyed. Before getting
destroyed, it removes itself from the window's drag-drop client. So when the
X11DragContext is being destroyed and tries to look up the drag-drop client, it
always fails, and so the platform-event dispatcher remains the dispatcher-list
of the platform-event source. This ends up causing a crash after the drag. So
unconditionally remove the dispatcher from the event-source upto destruction.

BUG=360477
R=erg@chromium.org

Review URL: https://codereview.chromium.org/227633003