linux-aura: Prepend 'http://' to selection-clipboard from omnibox selection.
Let the views::TextfieldController know about the clipboard that the textfield
content was copied into in OnAfterCutOrCopy(), so the controller can prepend
'http://' to the correct clipboard content (when appropriate). This fixes
pasting from the selection-clipboard on linux-aura after selecting the URL in
the omnibox.

BUG=351716
R=msw@chromium.org, sky@chromium.org

Review URL: https://codereview.chromium.org/227363003