[GCM] Fix crash due to connection retry while logging in.
It was possible for a backoff retry to attempt while a previous attempt
was in the middle of logging in. This could result in resetting the socket,
which would lead a crash later if the login failed. Protect against this by
ignoring backoff retries if we know we're logging in.

BUG=360680

Review URL: https://codereview.chromium.org/227513006