Generate the iOS policy Plist example when building on Mac.
This change triggers the ios_plist generator in grit when building
on the mac.

This change also lets the 'dict' policy type map to any JSON structure
(currently a dictionary or an array), so that the ManagedBookmarks
policy can have correct examples generated by the templates
generators.

BUG=360508

Review URL: https://codereview.chromium.org/225303013