Do not retain zombie CGL contexts
There are reports of failures when releasing the previous_context_ in
ScopedCGLSetCurrentContext's destructor. The exact reason is unknown,
but test applications have shown that it is possible for the current
context to have a reference count of zero (in which case, the retain
before setting the new context current does nothing, and the release
after the new context is not current anymore may crash).

If the previous context has a zero reference count, do not retain and
release it, rather, set the context to NULL after the call completes.

BUG=358473

Review URL: https://codereview.chromium.org/224723021