linux_aura: Fix unsafe thread usage crash in KDE select file dialogs.
Previously, we just passed pointers to GtkWidgets around in the gtk port
version of this code. In both GTK and aura, we were unsafely accessing
the same parents set on two different threads.

Fixed both issues. Now we pass XIDs to the file thread and only access
the |parents_| set on the UI thread.

BUG=360871

Review URL: https://codereview.chromium.org/228713005