metrics: Add UpdateEngine.SuccessfulUpdate.* metrics.
We've rewritten the metrics reported by the ChromeOS update_engine.
It's done in a way so the new metrics (UpdateEngine.*) will live
alongside the with the existing metrics (Installer.*). We plan to
eventually remove the old metrics.

This CL adds metrics in the UpdateEngine.SuccessfulUpdate.*
namespace. Many of these new metrics correspond - but are not 100%
identical - to the existing metrics in the Installer.* namespace.

BUG=chromium:355745
TEST=Validated the xml, pretty print succeeded.
R=asvitkine@chromium.org, avakulenko@chromium.org, deymo@chromium.org, isherman@chromium.org, jar@chromium.org, sosa@chromium.org

Review URL: https://codereview.chromium.org/227663004