Refactor PPAPI GLibc related tests.
This CL refactors PPAPI glibc tests to prepare adding non-SFI tests
in simpler form. As we need to enable or disable glibc tests and
non-SFI tests independently based on architecture, it would be verbose
to have single if-def chain to define TEST_PPAPI_NACL etc.
This CL tries to simplify it by using MAYBE_GLIBC, which is already
used widely, with a plan to introduce similar MAYBE_NONSFI macro
for non-SFI test in the next CL.

TEST=Ran trybots.
BUG=358425

Review URL: https://codereview.chromium.org/228943002