Improve GN public header file checking
This makes the header checker and include iterator work from InputFiles (which basically just hold the file buffer) rather than just raw strings. This allows us to reference these files from Err.

Some extra line/char tracking is now in the include iterator so we can actually quote from and annotate the place in the source file where the bad include is, which looks much nicer than "the file blah included blah". It also makes it possible to write much shorter error messages that still make sense.

This adds visibility checking as previously documented in the help for "public" and a unit test for that.

This updates the documentation for "public" which was wrong (it referred to patterns which was not used).

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/229423002