Set the default touch scroll mode back to touchcancel
Touch move absorption appears to be having too large of a web compatibility
impact, and there is debate about what the best fix is.  So for now go
back to the behavior we've always shipped in the past - sending touchcancel
on the start of scroll.

BUG=360614

Review URL: https://codereview.chromium.org/230733005