cc: Fix invalid tiling state after lost context
After a lost context, it used to be the case that a tree wouldn't need
update draw properties (implying no ManageTilings prior to AppendQuads
or MarkVisibleResourcesAsRequired).  This meant that if one layer added
a tiling and synced it to its twin, that the tiling state would not be
valid.  This synced tiling would be the first tiling, but the resolution
wouldn't be set, causing the tiling state to be bogus.

The way to fix this is to simply set_needs_update_draw_properties when a
PictureLayerImpl loses its resources.  This will force tilings to be
created and resolutions set correctly.

BUG=358350

Review URL: https://codereview.chromium.org/229283004