libevent: fix race in the common case.
There's a small race in the libevent code: it tests whether the monotonic clock
works and, if so, uses it and skips time-warp detection. However, the flag is a
static and thus causes a race.

We don't want to lose the ability to skip time-warp detection when the system
has a monotonic clock because we expect the vast majority of systems to have
it. But I'm not quite confident enough to remove support for systems without it
yet.

Thus this patch moves the race so that it only triggers on old systems that
don't have a monotonic clock. It's still there because we don't have easy
access to a "once" object this far down and we don't want to add a static init
function. But it will stop triggering the detectors and only arise on very old
systems.

BUG=360449
R=jln@chromium.org

Review URL: https://codereview.chromium.org/231493002