GN: Don't write duplicate phony rules.
In GN it's possible to have targets with the same short name as long as the files in the root dir don't collide. This is easiest to get with source_sets, where there is no output file and you don't have to worry about naming at all.

But currently Ninja will complain about duplicate rules since we write "short" phony rules for each target. This patch skips the phony rule shortcuts for targets with duplicate names. The places where I was seeing the collision problems was in the "test_support" type target where there is little reason to use the short name anyway.

Review URL: https://codereview.chromium.org/232063002