Cast: Use 2 threads for encoding on capable systems
We would like to use more than 1 thread for encoding to improve
performance. We do this only on a higher end system. The criteria is to
have at least 3 logical processors. This has worked well for WebRTC and
the same condition is applied here.

BUG=360966

Review URL: https://codereview.chromium.org/228313002