Mac: Recreate browser-side GL contexts on GPU switch
When a GPU switch is detected, poison all browser-side GL contexts,
which will cause them to be recreated.

Make CompositorSwapBuffers destroy and recreate the GL context
if it has been poisoned. Ensure that the CALayer for the old content
remain in place until the new layer is in place and correctly sized,
to avoid flickering.

BUG=361493

Review URL: https://codereview.chromium.org/231863002