Fix for often broken fling scrolls on Aura
Because the Animate() method was recently changed to pass the current
frame's tick time in r259735, getting the diff between it and the fling's
start time would often result in a negative time delta. This caused the
fling to report no scrolling and the scroll would stop. We still want to
start flinging immediately where possible so instead of letting Animate()
always set the startTime and losing a frame, we do it only for a
non-positive time delta.

This does not affect Android since its gesture fling ignores the passed
delta and uses gfx::FrameTime::Now().

BUG=360633

Review URL: https://codereview.chromium.org/232683003