Leak detector on build bots
This CL introduces the leak detector as chromium.webkit bots. The only
difference between these new bots and the existing WebKit bots is that
--enable-leak-detection is added to the run_webkit_test.py to execute
leak detection on layout tests.

The new bots are not tree closers for the duration.

We suppose that leaks should be treated as red errors, and the other
errors should be ignored like ASAN bots.

BUG=359024

Review URL: https://codereview.chromium.org/224963004