Do not call GetLastUsedProfile() when gathering extension data.
We should not call GetLastUsedProfile() in UMA code that gathers the
list of installed extensions, as that GetLastUsedProfile will create a
profile if one does not yet exist.  Use a more verbose alternative that
will simply return NULL if a profile is not yet available.

This should fix an observed crash on Windows.

BUG=360467

Review URL: https://codereview.chromium.org/235563003