Remove PageDistiller from DOM Distiller.
The PageDistiller adds an extra layer between the Distiller and the
underlying code which in fact distills the content.

This CL simplifies the distillation process by removing this class and
the delegate used for communcation. More of the logic is now moved into
the DistillerPage class.

BUG=361939

Review URL: https://codereview.chromium.org/231933005