PasswordManagerPresenter: early return instead of a DCHECK
Methods in the presenter which are called as a result of an API call from the chrome://settings/passwords renderer currently only have DCHECKS for verifying the bounds of indices into password lists.

If an out-of-bound index is passed from a compromised renderer, this would crash the browser. This CL therefore replaces the DCHECKS with an early return + a NOTREACHED (to keep the Debug build crashes informative).

R=markusheintz@chromium.org
BUG=362054

Review URL: https://codereview.chromium.org/234773002