Don't send two OnResponseCompleteds in ResourceLoader.
If the handler returns false in OnReadCompleted, we call OnResponseCompleted
twice. Add a test to ensure we don't do that. In addition, add a test for
deferring an EOF; this currently relies on URLRequest::Read being callable
after an EOF. The test will ensure we remove that assumption if it stops being
true.

BUG=361830

Review URL: https://codereview.chromium.org/231993003