Move HistoryController to content/renderer/
Currently, all history state updates are initiated from content/renderer, but are implemented in blink's Source/core/page. This moves the HistoryController class to content/renderer/, hangs it off of RenderViewImpl, and updates it to use chromium types.

I tried to change as little as possible in the logic of the code compared to src.chromium.org/viewvc/blink/trunk/Source/core/page/HistoryController.cpp.

This is step 4 of 5 in http://crbug.com/357327

BUG=357327

Review URL: https://codereview.chromium.org/229283002