Stop using v8::Persistent in ObjectBackedNativeHandler
When r261795 switched from storing UnsafePersistents to using the new
PersistentValueVector, the allocation of a v8::Persistent was erroneously
left in the code, and would never be disposed. As it happens, that Persistent
isn't needed at all, so avoiding the memory leak is just a matter of never
allocating it.

R=jochen@chromium.org
BUG=363708

Review URL: https://codereview.chromium.org/239253003