Update ServiceWorkerProviderHost when the pending version changes
When ServiceWorkerRegisterJob changes a registration's pending version, iterate
over all existing ServiceWorkerProviderHost and update their pending version,
if they are in-scope for the registration. This will enable documents to have
an up-to-date navigator.serviceWorker.pending object, but that is not yet
implemented (and will probably be renamed to next or installing).

BUG=361565

Review URL: https://codereview.chromium.org/238823002