Revert "Revert of Only create webrtc MediaStreams when added to a PeerConnection https://codereview.chromium.org/227163007/
This reverts commit caa70d180ff72cbf5f5ffd542b97de86794d5e00.

The reason for the original revert is fixed in https://codereview.chromium.org/237433003/.

Original CL desription:
This cl adds a new class WebRtcMediaStreamAdapter. The responsibility of the class is to create and own a representation of a webrtc MediaStream that can be added and removed from a PeerConnection.
An instance of WebRtcMediaStreamAdapter is created when a MediaStream is added to RTCPeerConnection object.

The purpose is to clean up the the webrtc specific use of MediaStreams.

TBR= ronghuawu@chromium.org
BUG=323223

Review URL: https://codereview.chromium.org/235503011