CoreAnimation: Fix text anti-aliasing in download shelf
Make the download shelf buttons opaque, and have them use the download
shelf's draw function to draw.

Add a helper function to look up an ancestor view based on the view ID.
This is particularly useful because the download shelf is five views
above the download shelf button view.

Update the bookmark button view function to refer to the toolbar view
as the download shelf toolbar view (which does not have a view ID
associated with it).

BUG=363750

Review URL: https://codereview.chromium.org/240093006