Update clang-format binaries & scripts to revision 206068
I followed these instructions:
https://code.google.com/p/chromium/wiki/UpdatingClangFormatBinaries

STATISTICS from an analysis of the 3283 most recent clang-formattable
commits to Chrome (~1 month of activity):

== How well did the old
== version of clang-format do?

47.49% commits were completely clang-format "clean" according to the
old version. This metric is an indicator of either clang-format
matching the prevailing style (which is more likely to happen with
smaller CLs), or people using clang-format as part of their workflow.

Between 78.11% and 86.58% of all committed lines (depending on how you
count -- the lower bound is very conservative) were clang-format
"clean" according to the old version. This metric is an indicator of
the deviations between prevailing style and the auto-formatting of
clang-format.

== How different is the new version of clang-format from the old
== version?

4.09% of lines in all commits would have been formatted differently
between the new and old versions of the tool.

30.28% commits (regardless of whether they were clean) would have had
at least one line formatted differently by the new version of clang-
format compared to the old version.

== What do the differences look like, qualitatively?

- const, OVERRIDE, and { are best friends that like to hang out.
- Just say no to one-line function definitions
- Enum declarations are one-per-line
- Statically initialized structs and arrays look different (more vertical whitespace)

Examples of all the above are here: http://pastebin.com/Ue4jvmvc

BUG=351534,360647,347174

Review URL: https://codereview.chromium.org/225143008