DCHECK that listener thread != IO thread
This came up in code review, where somebody was creating a SyncChannel on the IO thread. That will easily lead to deadlocks.

For ChannelProxy, it seems OK to have them be the same thread so long as there is no Listener. We apparently already do this in one place:
https://code.google.com/p/chromium/codesearch#chromium/src/components/nacl/browser/nacl_process_host.cc&l=866

BUG=

Review URL: https://codereview.chromium.org/234253002