Disallow touch hit-testing along attached edges of panels
Introduce AttachedPanelWindowTargeter, a new derived 
class of EasyResizeWindowTargeter. When installed as 
the EventTargeter of a panel container, the extended 
touch hit-testing inset is set to 0 for the edges of 
panels adjacent to the shelf. This makes it 
significantly easier to correctly target shelf buttons 
with touch because shelf-adjacent edges of panels are
prevented from themselves being targets.

BUG=351348
TEST=WorkspaceControllerTest.WindowEdgeTouchHitTestPanel,PanelLayoutManagerTest.TouchHitTestPanel

Review URL: https://codereview.chromium.org/232703006