Moves 'customize' button to a one-item menu below the block button.
Creates a new class, SplitBlockButton, which is a constrained window button
with two cells instead of just one.  One cell is the button, the other is
the menu's cell.

In order to share drawing code and colors with the constrained window
button, functions have been made public for reuse.

BUG=None

Review URL: https://codereview.chromium.org/208673002