Add histograms for network selection related metrics.
Added histogram for sampling number of auto-connectable wifi services
available when auto-connect is initiated for wifi device. This metric
is reported when auto-connect is initiated for wifi device.

Added histogram for sampling number of BSSes available when attempt
to connect to a wifi service. This metric is reported when connection
attempt is initiated on a wifi service.

Added histogram for sampling number of services that are connected
to the currently connected network. This metric is reported when
connection is established for a service.

BUG=chromium:230884,chromium:359302
TEST=Start chrome, and browse to chrome://histograms to make sure
these histograms exist after auto-connect to a wifi service.
NOTRY=true

Review URL: https://codereview.chromium.org/240503011